use double for total_score column instead

+++ database/migrations/2025_01_22_190453_create_user_season_score_aggregates_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
         Schema::create('user_season_score_aggregates', function (Blueprint $table) {
             $table->bigInteger('user_id')->unsigned();
             $table->integer('season_id')->unsigned();
-            $table->float('total_score');
+            $table->double('total_score');
 
             $table->primary(['user_id', 'season_id']);
             $table->index('total_score');
